refactor: comment and improve internals
remove dead code
