+++ packages/nextjs-mf/src/index.ts
@@ -1,4 +1,4 @@
-import { NextFederationPlugin } from './plugins/NextFederationPlugin/NextFederationPlugin';
+import NextFederationPlugin from "./plugins/NextFederationPlugin";
 
 export { NextFederationPlugin };
 export default NextFederationPlugin;
