+++ packages/nextjs-mf/src/internal.ts
@@ -1,19 +1,25 @@
-import type { Compiler } from 'webpack';
-import type {
-  ModuleFederationPluginOptions,
-  SharedConfig,
-  SharedObject,
-} from '@module-federation/utilities';
-import { createDelegatedModule } from '@module-federation/utilities';
+import type { ModuleFederationPluginOptions, SharedConfig, SharedObject } from "@module-federation/utilities";
+import { createDelegatedModule } from "@module-federation/utilities";
 
-import path from 'path';
+import { isRequiredVersion } from "webpack/lib/sharing/utils";
+import { parseOptions } from "webpack/lib/container/options";
 
-import { isRequiredVersion } from 'webpack/lib/sharing/utils';
-import { parseOptions } from 'webpack/lib/container/options';
+/**
+ * @typedef SharedObject
+ * @type {object}
+ * @property {object} [key] - The key representing the shared object's package name.
+ * @property {boolean} key.singleton - Whether the shared object should be a singleton.
+ * @property {boolean} key.requiredVersion - Whether a specific version of the shared object is required.
+ * @property {boolean} key.eager - Whether the shared object should be eagerly loaded.
+ * @property {boolean} key.import - Whether the shared object should be imported or not.
+ */
 
-// the share scope we attach by default
-// in hosts we re-key them to prevent webpack moving the modules into their own chunks (cause eager error)
-// in remote these are marked as import:false as we always expect the host to prove them
+/**
+ * A default share scope attached to shared objects.
+ * For server, they are marked as import:false, as the host is always expected to provide them as externals.
+ *
+ * @type {SharedObject}
+ */
 export const DEFAULT_SHARE_SCOPE: SharedObject = {
   react: {
     singleton: true,
@@ -22,12 +28,6 @@ export const DEFAULT_SHARE_SCOPE: SharedObject = {
     import: false,
   },
   'react/jsx-runtime': {
-    singleton: true,
-    requiredVersion: false,
-    import: false,
-    eager: false,
-  },
-  'react/jsx-dev-runtime': {
     singleton: true,
     requiredVersion: false,
     import: false,
@@ -89,6 +89,13 @@ export const DEFAULT_SHARE_SCOPE: SharedObject = {
   },
 };
 
+/**
+ * A default share scope for the browser environment.
+ * It takes the DEFAULT_SHARE_SCOPE and sets eager to true for all entries.
+ * The module hoisting system relocates these modules into the right runtime and out of the remote
+ *
+ * @type {SharedObject}
+ */
 export const DEFAULT_SHARE_SCOPE_BROWSER: SharedObject = Object.entries(
   DEFAULT_SHARE_SCOPE
 ).reduce((acc, item) => {
@@ -99,63 +106,40 @@ export const DEFAULT_SHARE_SCOPE_BROWSER: SharedObject = Object.entries(
   return acc;
 }, {} as SharedObject);
 
-export const externalizedShares: SharedObject = Object.entries(
-  DEFAULT_SHARE_SCOPE
-).reduce((acc, item) => {
-  const [key, value] = item as [string, SharedConfig];
-
-  acc[key] = { ...value, import: false };
-
-  if (key === 'react/jsx-runtime') {
-    delete (acc[key] as SharedConfig).import;
-  }
-
-  return acc;
-}, {} as SharedObject);
-
-// determine output base path, derives .next folder location
-export const getOutputPath = (compiler: Compiler) => {
-  const isServer = compiler.options.target !== 'client';
-  let outputPath: string | string[] | undefined =
-    compiler.options.output.path?.split(path.sep);
-
-  const foundIndex = outputPath?.lastIndexOf(isServer ? 'server' : 'static');
-
-  outputPath = outputPath
-    ?.slice(0, foundIndex && foundIndex > 0 ? foundIndex : outputPath.length)
-    .join(path.sep);
-
-  return outputPath as string;
+/**
+ * Checks if the remote value is an internal or promise delegate module reference.
+ *
+ * @param {string} value - The remote value to check.
+ * @returns {boolean} - True if the value is an internal or promise delegate module reference, false otherwise.
+ */
+const isInternalOrPromise = (value: string): boolean => {
+  return value.startsWith('internal ') || value.startsWith('promise ');
 };
 
-export const removePlugins = [
-  'NextJsRequireCacheHotReloader',
-  'BuildManifestPlugin',
-  'WellKnownErrorsPlugin',
-  'WebpackBuildEventsPlugin',
-  'HotModuleReplacementPlugin',
-  'NextMiniCssExtractPlugin',
-  'NextFederationPlugin',
-  'CopyFilePlugin',
-  'ProfilingPlugin',
-  'DropClientPage',
-  'ReactFreshWebpackPlugin',
-  'NextMedusaPlugin',
-];
+/**
+ * Checks if the remote value is using the standard remote syntax.
+ *
+ * @param {string} value - The remote value to check.
+ * @returns {boolean} - True if the value is using the standard remote syntax, false otherwise.
+ */
+const isStandardRemoteSyntax = (value: string): boolean => {
+  return value.includes('@');
+};
 
-/*
- This code is doing the following It's iterating over all remotes and checking if
- they are using a custom promise template or not If it's a custom promise template
- we're parsing the remote syntax to get the module name and version number
-  */
-export const parseRemotes = (remotes: Record<string, any>) =>
-  Object.entries(remotes).reduce((acc, [key, value]) => {
-    // check if user is passing a internal "delegate module" reference
-    if (value.startsWith('internal ') || value.startsWith('promise ')) {
+/**
+ * Parses the remotes object, checking if they are using a custom promise template or not.
+ * If it's a custom promise template, the remote syntax is parsed to get the module name and version number.
+ *
+ * @param {Record<string, any>} remotes - The remotes object to be parsed.
+ * @returns {Record<string, string>} - The parsed remotes object.
+ */
+export const parseRemotes = (remotes: Record<string, any>): Record<string, string> => {
+  return Object.entries(remotes).reduce((acc, [key, value]) => {
+    if (isInternalOrPromise(value)) {
       return { ...acc, [key]: value };
     }
-    // check if user is passing standard remote syntax, if so use default delegate
-    if (value.includes('@')) {
+
+    if (isStandardRemoteSyntax(value)) {
       return {
         ...acc,
         [key]: createDelegatedModule(require.resolve('./default-delegate.js'), {
@@ -163,14 +147,30 @@ export const parseRemotes = (remotes: Record<string, any>) =>
         }),
       };
     }
-    // return standard template otherwise
+
     return { ...acc, [key]: value };
   }, {} as Record<string, string>);
+};
 
-export const getDelegates = (remotes: Record<string, any>) => {
+/**
+ * Checks if the remote value is an internal delegate module reference.
+ *
+ * @param {string} value - The remote value to check.
+ * @returns {boolean} - True if the value is an internal delegate module reference, false otherwise.
+ */
+const isInternalDelegate = (value: string): boolean => {
+  return value.startsWith('internal ');
+};
+
+/**
+ * Gets the delegate modules from the remotes object.
+ *
+ * @param {Record<string, any>} remotes - The remotes object containing delegates.
+ * @returns {Record<string, string>} - The delegate modules from the remotes object.
+ */
+export const getDelegates = (remotes: Record<string, any>): Record<string, string> => {
   return Object.entries(remotes).reduce((acc, [key, value]) => {
-    // check if user is passing a internal "delegate module" reference
-    if (value.startsWith('internal ')) {
+    if (isInternalDelegate(value)) {
       return { ...acc, [key]: value };
     }
 
@@ -178,31 +178,41 @@ export const getDelegates = (remotes: Record<string, any>) => {
   }, {} as Record<string, string>);
 };
 
-/*
- This code is parsing the options shared object and creating a new object with all
- of the shared configs Then it is iterating over each key in this new object and
- assigning them to an array that will be returned by this function This array contains
- objects that are used as values for the shared property of Module Federation Plugin
- Options The first thing we do here is check if the item passed into shared was a
- string or not if it\'s an array If it wasn\'t then throw an error because there should
- only be strings in there Otherwise continue on with our code below
-  */
+/**
+ * Validates the shared item type and returns the shared config object based on the item and key.
+ *
+ * @param {string} item - The shared item.
+ * @param {string} key - The key for the shared item.
+ * @returns {object} - The shared config object.
+ * @throws {Error} - If the item type is not a string.
+ */
+const getSharedConfig = (item: string, key: string) => {
+  if (typeof item !== 'string') {
+    throw new Error('Unexpected array in shared');
+  }
+
+  return item === key || !isRequiredVersion(item)
+    ? {
+      import: item,
+    }
+    : {
+      import: key,
+      requiredVersion: item,
+    };
+};
+
+/**
+ * Parses the share options in ModuleFederationPluginOptions and creates a new object with all
+ * of the shared configs. This object is then used as the value for the shared property
+ * of Module Federation Plugin Options.
+ *
+ * @param {ModuleFederationPluginOptions} options - The ModuleFederationPluginOptions object.
+ * @returns {Record<string, SharedConfig>} - The parsed share options object.
+ */
 const parseShareOptions = (options: ModuleFederationPluginOptions) => {
   const sharedOptions: [string, SharedConfig][] = parseOptions(
     options.shared,
-    (item: string, key: string) => {
-      if (typeof item !== 'string')
-        throw new Error('Unexpected array in shared');
-
-      return item === key || !isRequiredVersion(item)
-        ? {
-            import: item,
-          }
-        : {
-            import: key,
-            requiredVersion: item,
-          };
-    },
+    getSharedConfig,
     (item: any) => item
   );
 
@@ -221,13 +231,27 @@ const parseShareOptions = (options: ModuleFederationPluginOptions) => {
   }, {} as Record<string, SharedConfig>);
 };
 
-export const toDisplayErrors = (err: Error[]) =>
-  err
-    .map((error) => {
-      let { message } = error;
-      if (error.stack) {
-        message += `\n${error.stack}`;
-      }
-      return message;
-    })
-    .join('\n');
+/**
+ * Formats an error object into a displayable string.
+ *
+ * @param {Error} error - The error object to format.
+ * @returns {string} - The formatted error string.
+ */
+const formatError = (error: Error): string => {
+  let { message } = error;
+  if (error.stack) {
+    message += `\n${error.stack}`;
+  }
+  return message;
+};
+
+/**
+ * Converts an array of Error objects into a single string with formatted error messages,
+ * separated by newline characters.
+ *
+ * @param {Error[]} err - The array of Error objects.
+ * @returns {string} - The combined formatted error messages as a single string.
+ */
+export const toDisplayErrors = (err: Error[]): string => {
+  return err.map(formatError).join('\n');
+};
