+++ packages/utilities/src/utils/common.ts
@@ -2,13 +2,13 @@
 
 import type {
   AsyncContainer,
-  Remotes,
-  RuntimeRemotesMap,
-  RuntimeRemote,
-  WebpackRemoteContainer,
-  RemoteData,
   GetModuleOptions,
-} from '../types';
+  RemoteData,
+  Remotes,
+  RuntimeRemote,
+  RuntimeRemotesMap,
+  WebpackRemoteContainer
+} from "../types";
 
 type RemoteVars = Record<
   string,
@@ -311,7 +311,7 @@ const createContainerSharingScope = (
         //   `container.init` does not return a Promise, and here we do not call `then` on it.
         // But according to [docs](https://webpack.js.org/concepts/module-federation/#dynamic-remote-containers)
         //   it must be async.
-        // The problem may be in Proxy in NextFederationPlugin.js.
+        // The problem may be in Proxy in Index.js.
         //   or maybe a bug in the webpack itself - instead of returning rejected promise it just throws an error.
         // But now everything works properly and we keep this code as is.
         container.init(__webpack_share_scopes__['default'] as any);
